requirements.txt
# Core LLM Libraries
transformers>=4.41.2
accelerate>=0.31.0
openai>=1.0.0
langchain>=0.1.0
google-search-results>=2.4.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
fastapi>=0.116.1
uvicorn>=0.24.0
